now texture

+++ include/glm/gtc/random.inl
@@ -22,7 +22,7 @@ namespace detail
 		GLM_FUNC_QUALIFIER static vec<1, uint8, P> call()
 		{
 			return vec<1, uint8, P>(
-				static_cast<uint8>(std::rand() % std::numeric_limits<uint8>::max()));
+				std::rand() % std::numeric_limits<uint8>::max());
 		}
 	};
